Added faq page to route-pages. * Thanks to Devoder we have an Es translation!   * This will allow reference on switching route-pages languages.
