+++ src/modules/route-pages/element.component.tsx
@@ -1,6 +1,6 @@
 import { Common } from '@modules/common';
 import { Route, Routes } from 'react-router-dom';
-import { ContactPage, CookiesPage, PrivacyPage } from './pages';
+import { ContactPage, CookiesPage, FaqPage, PrivacyPage } from './pages';
 
 export const Element: React.FC = () => {
     return (
@@ -8,6 +8,7 @@ export const Element: React.FC = () => {
             <Route path='contact' element={<ContactPage />} />
             <Route path='cookies' element={<CookiesPage />} />
             <Route path='privacy' element={<PrivacyPage />} />
+            <Route path='faq' element={<FaqPage />} />
             <Route path='*' index element={<Common.NotFoundPage />} />
         </Routes>
     );
