+++ src/modules/route-pages/pages/faq.page.enEs.md
@@ -1,0 +1,64 @@
+# Preguntas Frecuentes
+
+Bienvenido a Diablo4.trading. Esta es una pequeña sección de preguntas frecuentes donde se responderán preguntas básicas. Por favor, consulta nuestra guía completa [aquí](https://discord.com/invite/diablo4).
+
+---
+
+## General
+
+No puedes comerciar, publicar servicios o utilizar el sitio web sin iniciar sesión en una cuenta de Discord.
+
+Todas las reglas del Discord del Santuario son aplicables al sitio web.
+
+Cualquier abuso hacia el sitio web, usuarios, sistema de avales y demás, resultará en que tu cuenta y cuentas afiliadas sean incluidas en la lista negra.
+
+---
+
+## Inicio de Sesión
+
+**Cómo Acceder al Sistema de Comercio**
+- P: ¿Cómo accedo al sistema de comercio?
+- R: Conecta tu cuenta de Discord haciendo clic en el botón "Iniciar Sesión" en la parte superior derecha del sitio web para ingresar al sitio.
+
+---
+
+## Sistema de Comercio
+
+- P: ¿Para qué sirve el sistema de comercio?
+- R: El sistema de comercio es una plataforma de código abierto creada por el equipo de desarrolladores del Discord del Santuario de Diablo 4 para que los jugadores compren, vendan e intercambien objetos del juego de forma segura.
+
+- P. ¿Cómo puedo comprar objetos?
+- R. Navega por la página de listados, selecciona un objeto y sigue las instrucciones proporcionadas para adquirir el objeto.
+
+- P. ¿Cómo puedo vender objetos?
+- R. Puedes vender objetos haciendo clic en la página "Comercio", seleccionando "Mis Listados" y luego haciendo clic en "Crear Nuevo Listado".
+
+---
+
+## Sistema de Servicios
+
+- P. ¿Qué es el sistema de servicios?
+- R. El sistema de servicios permite a los jugadores ofrecer y solicitar servicios dentro del juego entre sí, como ayuda en Mazmorras de Pesadilla, renombre y nivelación.
+
+- P. ¿Cómo compro servicios de alguien?
+- R. Navega por los listados de servicios, elige un vendedor y sigue las instrucciones proporcionadas para adquirir el servicio.
+
+---
+
+## Sistema de Informes
+
+- P. ¿Qué es el sistema de informes?
+- R. El sistema de informes es un sistema en marcha para tomar medidas de las pautas comunitarias contra los usuarios que rompen las reglas.
+
+- P. ¿Cómo informo a un usuario?
+- R. Puedes informar a un usuario seleccionando en la página del usuario, haciendo clic en el botón "informar" y luego siguiendo las instrucciones.
+
+---
+
+## Sistema de Avales
+
+- P. ¿Qué es el sistema de avales?
+- R. El sistema de avales es un sistema desarrollado por el personal técnico del Santuario para permitir que los usuarios obtengan un estatus positivo o negativo.
+
+- P. ¿Cómo puedo avalar a alguien?
+- R. Puedes avalar a alguien haciendo clic en el perfil del usuario, seleccionando un aval positivo o negativo, dando una razón y un pequeño texto para explicar por qué.
