+++ src/modules/route-pages/pages/index.ts
@@ -1,3 +1,4 @@
 export * from './contact.page';
 export * from './cookies.page';
+export * from './faq.page';
 export * from './privacy.page';
